+++ Makie/test/runtests.jl
@@ -61,6 +61,10 @@ end
         include("events_isolation.jl")
     end
 
+    @testset "GUI" begin
+        include("gui.jl")
+    end
+
     @testset "Conversion & Projection Pipeline" begin
         # TODO: move some things in here
         include("conversions/convert_arguments.jl")